Removes old test, adds MockRetriever to test file and test uses document_store
